requirement.txt
requests
pymysql
pycryptodome
retry